feat: update the webapi route of the trace facade
